ai/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
httpx>=0.24.0
python-dotenv>=1.0.0
